Style fixes for mobile devices.
It turns out that glob-order is not predictable, and we need to
include our CSS (and JS) in a particular order. So I explicitly state
the order of the included files in the Pipeline settings.

This isn't quite repeating myself, as the order is information.

I also added some media-query settings for the top padding, as mobile
devices don't need it.